Uploaded the nlp_grading_system script
This version only validates a CSV file (Improvements will be made in the nearest future), checks for missing columns, checks for missing values, and also checks for duplicate records. It should refresh the page and clear uploaded files too (it only refreshes the page), but it seems *streamlit* has scrapped some of the functions for that. I'll try other ways.